pyproject.toml
[project]
name = "rau-backend"
version = "0.1.0"
description = "RAU backend (FastAPI + Postgres)."
requires-python = ">=3.11"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "python-dotenv>=1.0",
  "pydantic>=2.7",
  "pydantic-settings>=2.5",
  "pyjwt>=2.9",
  "passlib[bcrypt]>=1.7"
]

[tool.uvicorn]
factory = false
host = "0.0.0.0"
port = 8000
reload = true
